refactor: replace exceptions by null
It's a conscious design decision to not throw, but instead leave it to
the consumer to decide how to handle those situations. E.g. if an
index is out of bounds, the consumer could still recover when properly
processing null (just like nil in Clojure) instead of pulling the rug
out from under him with an exception.