feat: introduce alternative lazy-seq
This implementation is closer to Clojure's version
which comes with the caveat that the Seq can be
empty. This requires additional checks, which were
not necessary previously as only Nil could ever be
empty. However, the advantage is that any type of
seq can be wrapped in the new lazy seq which
allows for a number of optimizations. I am of the
opinion that this justifies the caveat as it may
lead to substantial performance improvements down
the road - if I'm able to replace the old
implementation by this new one.