refactor: remove caching
Seqs are already cached; accessing its list features should not waste
twice as much memory for caching the same data in list-form. Instead,
we accept the trade-off of building a list from the seq every time it
is accessed. It is to be expected that seq-features are much more
frequently used than list-features. If someone needs a list they should
not use a seq in the first place.