refactor: remove force()
Since we now fully implement List (via AList), we don't need a
recursive force() any longer. A list from a seq will be created
linearly by walking over the items, realizing it on the way (or running
forever if infinite). This might be slower for large seqs, but prevents
StackOverflowErrors.